fix: review comments - move dhis specific entities to data layer
