+++ src/domain/repositories/ModuleRepository.ts
@@ -1,6 +1,6 @@
-import { Id } from "@eyeseetea/d2-api";
 import { FutureData } from "../../data/api-futures";
 import { AMRSurveyModule } from "../entities/AMRSurveyModule";
+import { Id } from "../entities/Ref";
 
 export interface ModuleRepository {
     getAll(): FutureData<AMRSurveyModule[]>;
